[core] Move QRCodeLoginListener to `.auth`
