+++ mirai-core-api/src/commonMain/kotlin/utils/LoginSolver.kt
@@ -13,6 +13,7 @@ package net.mamoe.mirai.utils
 
 import me.him188.kotlin.jvm.blocking.bridge.JvmBlockingBridge
 import net.mamoe.mirai.Bot
+import net.mamoe.mirai.auth.QRCodeLoginListener
 import net.mamoe.mirai.network.LoginFailedException
 import net.mamoe.mirai.network.RetryLaterException
 import net.mamoe.mirai.network.UnsupportedSmsLoginException
@@ -27,21 +28,6 @@ import kotlin.jvm.JvmName
  * @see BotConfiguration.loginSolver
  */
 public abstract class LoginSolver {
-    /**
-     * 使用二维码登录时获取的二维码图片大小.
-     */
-    public open val qrCodeSize: Int = 3
-
-    /**
-     * 使用二维码登录时获取的二维码边框宽度.
-     */
-    public open val qrCodeMargin: Int = 4
-
-    /**
-     * 使用二维码登录时获取的二维码校正等级，必须为 1-3 之间.
-     */
-    public open val qrCodeEcLevel: Int = 2
-
     /**
      * 处理图片验证码, 返回图片验证码内容.
      *
@@ -146,56 +132,6 @@ public abstract class LoginSolver {
     public open suspend fun onSolveUnsafeDeviceLoginVerify(bot: Bot, url: String): String? {
         // This function was abstract, open since 2.13.0
         throw UnsupportedSmsLoginException("This login session requires device verification, but current LoginSolver($this) does not support it. Please override `LoginSolver.onSolveDeviceVerification`.")
-    }
-
-    /**
-     * 二维码扫描登录监听器
-     * @since 2.15
-     */
-    public interface QRCodeLoginListener {
-        /**
-         * 从服务器获取二维码时调用，在下级显示二维码并扫描.
-         */
-        public fun onFetchQRCode(bot: Bot, data: ByteArray)
-
-        /**
-         * 当二维码状态变化时调用.
-         * @see State
-         */
-        public fun onStatusChanged(bot: Bot, state: State)
-
-        public enum class State {
-            /**
-             * 等待扫描中，请在此阶段请扫描二维码.
-             * @see QRCodeLoginListener.onFetchQRCode
-             */
-            WAITING_FOR_SCAN,
-
-            /**
-             * 二维码已扫描，等待扫描端确认登录.
-             */
-            WAITING_FOR_CONFIRM,
-
-            /**
-             * 扫描后取消了确认.
-             */
-            CANCELLED,
-
-            /**
-             * 二维码超时，必须重新获取二维码.
-             */
-            TIMEOUT,
-
-            /**
-             * 二维码已确认，将会继续登录.
-             */
-            CONFIRMED,
-
-            /**
-             * 默认状态，在登录前通常为此状态.
-             */
-            DEFAULT,
-        }
     }
 
     public companion object {
