+++ mirai-core/src/commonMain/kotlin/network/components/QRCodeLoginProcessor.kt
@@ -11,13 +11,13 @@ package net.mamoe.mirai.internal.network.components
 
 import kotlinx.atomicfu.atomic
 import kotlinx.coroutines.delay
+import net.mamoe.mirai.auth.QRCodeLoginListener
 import net.mamoe.mirai.internal.network.QQAndroidClient
 import net.mamoe.mirai.internal.network.QRCodeLoginData
 import net.mamoe.mirai.internal.network.component.ComponentKey
 import net.mamoe.mirai.internal.network.handler.NetworkHandler
 import net.mamoe.mirai.internal.network.protocol.packet.login.WtLogin
 import net.mamoe.mirai.internal.utils.MiraiProtocolInternal.Companion.asInternal
-import net.mamoe.mirai.utils.LoginSolver
 import net.mamoe.mirai.utils.MiraiLogger
 import net.mamoe.mirai.utils.debug
 
@@ -66,14 +66,14 @@ internal class QRCodeLoginProcessorPreLoaded(
 }
 
 internal class QRCodeLoginProcessorImpl(
-    private val qrCodeLoginListener: LoginSolver.QRCodeLoginListener,
+    private val qrCodeLoginListener: QRCodeLoginListener,
     private val size: Int,
     private val margin: Int,
     private val ecLevel: Int,
     private val logger: MiraiLogger,
 ) : QRCodeLoginProcessor {
 
-    private var state = atomic(LoginSolver.QRCodeLoginListener.State.DEFAULT)
+    private var state = atomic(QRCodeLoginListener.State.DEFAULT)
 
     private suspend fun requestQRCode(
         handler: NetworkHandler,
@@ -133,24 +133,24 @@ internal class QRCodeLoginProcessorImpl(
         }
     }
 
-    private fun WtLogin.TransEmp.Response.mapProtocolState(): LoginSolver.QRCodeLoginListener.State {
+    private fun WtLogin.TransEmp.Response.mapProtocolState(): QRCodeLoginListener.State {
         return when (this) {
             is WtLogin.TransEmp.Response.QRCodeStatus -> when (this.state) {
                 WtLogin.TransEmp.Response.QRCodeStatus.State.WAITING_FOR_SCAN ->
-                    LoginSolver.QRCodeLoginListener.State.WAITING_FOR_SCAN
+                    QRCodeLoginListener.State.WAITING_FOR_SCAN
 
                 WtLogin.TransEmp.Response.QRCodeStatus.State.WAITING_FOR_CONFIRM ->
-                    LoginSolver.QRCodeLoginListener.State.WAITING_FOR_CONFIRM
+                    QRCodeLoginListener.State.WAITING_FOR_CONFIRM
 
                 WtLogin.TransEmp.Response.QRCodeStatus.State.CANCELLED ->
-                    LoginSolver.QRCodeLoginListener.State.CANCELLED
+                    QRCodeLoginListener.State.CANCELLED
 
                 WtLogin.TransEmp.Response.QRCodeStatus.State.TIMEOUT ->
-                    LoginSolver.QRCodeLoginListener.State.TIMEOUT
+                    QRCodeLoginListener.State.TIMEOUT
             }
 
             is WtLogin.TransEmp.Response.QRCodeConfirmed ->
-                LoginSolver.QRCodeLoginListener.State.CONFIRMED
+                QRCodeLoginListener.State.CONFIRMED
 
             is WtLogin.TransEmp.Response.FetchQRCode ->
                 error("$this cannot be mapped to listener state.")
